api/requirements.txt
accelerate==0.27.2
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.0.5
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
executing==0.8.3
fastapi==0.110.0
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
IProgress==0.4
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.2
jedi==0.18.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_widgets==3.0.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.12.0
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pip==23.3.1
platformdirs==3.10.0
prompt-toolkit==3.0.43
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
stack-data==0.2.0
starlette==0.36.3
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
tornado==6.3.3
tqdm==4.66.2
traitlets==5.7.1
transformers==4.38.1
triton==2.2.0
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
wcwidth==0.2.5
wheel==0.41.2
widgetsnbextension==4.0.10
